requirements.txt
requests
pandas
torch
langsmith
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain-core
pandas
gradio